fix(domino-royal): restore arena markup and route game cards to js entry
